# Conflicts: All resolved manually (with a lot of help from fork's amazing merge conflict tool)
#	Entities/ACDropShip.cpp
#	Entities/ACRocket.cpp
#	Entities/ACrab.cpp
#	Entities/ACraft.cpp
#	Entities/ADoor.cpp
#	Entities/AEmitter.cpp
#	Entities/AHuman.cpp
#	Entities/Actor.cpp
#	Entities/Arm.cpp
#	Entities/Attachable.cpp
#	Entities/HDFirearm.cpp
#	Entities/Leg.cpp
#	Entities/MOSRotating.cpp
#	Entities/Turret.cpp
#	System/ContentFile.cpp

+++ Entities/Actor.h
@@ -522,7 +522,7 @@ ClassInfoGetters
 //                  Actor.
 // Return value:    None.
 
-    void AddGold(float goldOz) { m_GoldCarried += ceilf(goldOz); m_GoldPicked = true; }
+    void AddGold(float goldOz) { m_GoldCarried += std::ceil(goldOz); m_GoldPicked = true; }
 */
 
 //////////////////////////////////////////////////////////////////////////////////////////
